Refactor structure and ui especially in mobile screen on Home, Credential History and Credential Details
